+++ soundProcessor/slowConverter.cpp
@@ -32,11 +32,11 @@ void slowConverter::whatAreYouDoing(FILE* fout) {
 	fwrite(info.data(), sizeof(char), info.size(), fout);
 }
 
-Stream slowConverter::convert(std::vector<std::string> streamFiles, std::vector<unsigned int> parameters, std::shared_ptr<std::string> outputFile = nullptr) {
+Stream slowConverter::convert(std::vector<std::string> streamFiles, std::vector<double> parameters, std::shared_ptr<std::string> outputFile = nullptr) {
 
-	unsigned int time_begin = 0;
-	unsigned int duration;
-	unsigned int ratio;
+	double time_begin = 0;
+	double duration;
+	double ratio;
 
 	std::string streamFile = streamFiles[0];
 	if (parameters.size() == 3) {
@@ -94,7 +94,7 @@ Stream slowConverter::convert(std::vector<std::string> streamFiles, std::vector<
 		throw std::runtime_error("Unavailable output file for slowing");
 	}
 
-	newStream.getHeader().get_subchunk3_size() = ratio * newStream.getHeader().get_subchunk3_size();
+	newStream.getHeader().get_subchunk3_size() = static_cast<size_t>(ratio * newStream.getHeader().get_subchunk3_size());
 	outputHeader outputHeader(fout, newStream.getHeader());
 	try {
 		outputHeader.output();
@@ -107,12 +107,12 @@ Stream slowConverter::convert(std::vector<std::string> streamFiles, std::vector<
 	writeBuffer writeBuff(BUFF_SIZE, fout, newStream.getData());
 
 	size_t data_size = (stream.getHeader().get_chunk_size() - stream.getData()) / 2;
-	size_t begin = time_begin * stream.getHeader().get_sample_rate();
+	size_t begin = static_cast<size_t>(time_begin * stream.getHeader().get_sample_rate());
 	if (begin > data_size) {
 		throw std::runtime_error("Unavailable argument of begin_time for slowing");
 	}
 
-	size_t end = (time_begin + duration) * stream.getHeader().get_sample_rate();
+	size_t end = static_cast<size_t>((time_begin + duration) * stream.getHeader().get_sample_rate());
 	if (end > data_size) {
 		throw std::runtime_error("Unavailable argument of duration for slowing");
 	}
