+++ soundProcessor/slowConverter.h
@@ -10,7 +10,7 @@ public:
 	slowConverter() = default;
 	~slowConverter() = default;
 
-	Stream convert(std::vector<std::string>, std::vector<unsigned int>, std::shared_ptr<std::string>) override;
+	Stream convert(std::vector<std::string>, std::vector<double>, std::shared_ptr<std::string>) override;
 	void whatAreYouDoing(FILE*) override;
 };
